+++ CHANGELOG.md
@@ -9,6 +9,7 @@ O versionamento deste projeto é aderente aos princípios de [Semantic Versionin
 ### Changed
 
 - Adicionada verificação no `<cps-input>` para que, ao usar `type` com os valores `button`, `checkbox`, `image`, `radio`, `range`, `reset`, `submit`, se instrua sobre a utilização dos componentes corretos.
+- Adiciona suporte ao `type` `hidden` no `<cps-input>`.
 
 ## 0.20.0 - 2025-06-24